requirements.txt
greenlet==1.1.2
playwright==1.20.1
pyee==8.1.0
python-dotenv==0.20.0
websockets==10.1
